rename frontend InstanceSegmentation files to SegmentationMask
