+++ lightly_studio_view/e2e/general/export-segmentation-masks.e2e-test.ts
