--- lightly_studio_view/src/lib/hooks/useInstanceSegmentationPreview/index.ts
@@ -1,1 +1,0 @@
-export { useSegmentationMaskPreview } from './useInstanceSegmentationPreview';
