+++ lightly_studio_view/src/lib/hooks/useSegmentationMaskBrush.ts
