+++ lightly_studio_view/src/lib/hooks/useSegmentationMaskPreview/brushDrawer.test.ts
