+++ lightly_studio_view/src/lib/hooks/useSegmentationMaskPreview/canvasMaskManager.ts
