+++ lightly_studio_view/src/lib/hooks/useSegmentationMaskPreview/index.ts
@@ -1,0 +1,1 @@
+export { useSegmentationMaskPreview } from './useSegmentationMaskPreview';
