+++ lightly_studio_view/src/lib/hooks/useSegmentationMaskPreview/previewRenderer.test.ts
